move keyboard logic to helper function
