+++ app/src/main/java/de/jrpie/android/launcher/ui/list/apps/ListFragmentApps.kt
@@ -12,17 +12,14 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import de.jrpie.android.launcher.R
 import de.jrpie.android.launcher.apps.AppFilter
 import de.jrpie.android.launcher.databinding.ListAppsBinding
 import de.jrpie.android.launcher.preferences.LauncherPreferences
 import de.jrpie.android.launcher.ui.UIObject
-import de.jrpie.android.launcher.ui.closeSoftKeyboard
+import de.jrpie.android.launcher.ui.applyKeyboardSettings
 import de.jrpie.android.launcher.ui.list.AbstractListActivity
 import de.jrpie.android.launcher.ui.list.AppListActivity
-import de.jrpie.android.launcher.ui.openSoftKeyboard
-import kotlin.math.absoluteValue
 
 
 /**
@@ -106,19 +103,6 @@ class ListFragmentApps : Fragment(), UIObject {
                     }
                 }
             adapter = appsRecyclerAdapter
-            if (LauncherPreferences.functionality().searchAutoCloseKeyboard()) {
-                addOnScrollListener(object : RecyclerView.OnScrollListener() {
-                    var totalDy: Int = 0
-                    override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                        totalDy += dy
-
-                        if (totalDy.absoluteValue > 100) {
-                            totalDy = 0
-                            closeSoftKeyboard(requireActivity())
-                        }
-                    }
-                })
-            }
         }
 
         binding.listAppsSearchview.setOnQueryTextListener(object :
@@ -178,11 +162,8 @@ class ListFragmentApps : Fragment(), UIObject {
             (activity as? AppListActivity)?.updateTitle()
         }
 
-        if (listActivity.intention == AbstractListActivity.Companion.Intention.VIEW
-            && LauncherPreferences.functionality().searchAutoOpenKeyboard()
-        ) {
-            binding.listAppsSearchview.openSoftKeyboard(requireContext())
-        }
+        applyKeyboardSettings(requireActivity(), binding.listAppsRview, binding.listAppsSearchview,
+        listActivity.intention == AbstractListActivity.Companion.Intention.VIEW)
     }
 
 }
